Include module docstring in index.rst (by the magic of automodule).
Update trie’s module docstring and include it in the index.rst file.
The docstring is mostly what README.rst file contains except for
Installation section.  Duplicating stuff is not good in general but
in this instance it allows stripping of the :class: formatting which
confused both GitHub and PiPI.