Disable sorting of child nodes
dict does not sort keys so there is no need for Trie to do that
either. This improves performance for iterating and traversing the
trie.  Old behaviour (i.e. sorting of child nodes) can be brought
back with enable_sorting method.

Since this is backwards incompatible change, incrementing major
version number is arguably the right thing to do.