Use sphinxcontrib.napoleon to parse Google-style docs
Add sphinxcontrib.napoleon to the conf.py and enable Google docstring
parsing so that Args, Returns, Rises etc. sections are parsed correctly.

To be able to build the documentation, sphinxcontrib-napoleon package
is now necessary.  It can be installed via:
	pip install sphinxcontrib-napoleon

I’m not super happy with how sphinxcontrib.napoleon formats the
documentation (the pink background is really jarring) but it’s
still better than Sphinx’s default of not parsing Google docs
properly.