Make shortest_prefix and longest_prefix falsy if no prefix found
Checking whether shortest_prefix and longest_prefix found a key
is a bit awkward.  It requires result tuple to be deconstructed
or its first element accessed.

To simplify things, change the methods to return falsy value in
those cases.  This allows them to be easily used in an ‘if’ or
‘while’ statement.  However, to still be consistent with result
type when prefix is found, keep it as a tuple.