Skip iterating over children when there aren't any
In case of StringTrie with sorting enabled, it was observed that filesystem paths that correspond to a directory when entered as a key in the Trie without any children lead to errors in the traverse method. This is due to the fact that such entries usually have a False value at the leaf nodes and thus, have no children.